app/models/formulaires.py
from flask_wtf import FlaskForm
from wtforms import StringField, SelectField, DateField, TextAreaField
from wtforms.validators import Optional

# test pour augmenter le dynamisme, a voir si je garde ou pas
"""Liste_faucons = [('', '-- Tous --')] + [
    ('4186361', 'Cleo'),
    ('4119953', 'Willy'),
    ('4186357', 'Roy'),
    ('41863610', 'Elmo'),
    ('4189880', 'Winnie'),
    ('BA12085', 'Rider'),
    ('BA12086', 'Peggy'),
    ('BA12100', 'Prince'),
    ('BA12620', 'Odin'),
    ('BA12839', 'Tofu'),
    ('BA12840', 'Pepper'),
    ('BA12841', 'Colossus'),
    ('BA3374', 'Ghost'),
    ('BA3382', 'Yoyo'),
    ('BA3396', 'Chico'),
    ('BA4010', 'Boomer'),
    ('BA8908', 'Claw'),
    ('BA8912', 'Sushi'),
    ('BA8916', 'Darcie'),
    ('BA8924', 'Sparrow'),
    ('BA8926', 'Sammy'),
    ('BA8977', 'Milo'),
    ('BA8982', 'Goliath'),
    ('BA8983', 'Hermes'),
    ('BA8984', 'Sky'),
    ('BA8985', 'Roxy'),
    ('BA8989', 'Aladdin'),
    ('BA8990', 'Frankie'),
    ('C12016', 'Buddy'),
    ('C12040', 'Joker'),
    ('C12046', 'Casper'),
    ('C12069', 'Apple'),
    ('C12442', 'Elvis'),
]"""

class Recherche(FlaskForm):
    code_faucon = SelectField("falcon_code", choices=[], validators=[Optional()])
    date_filtre  = DateField("Date", validators=[Optional()], format='%Y-%m-%d')
    place = StringField("space_label", validators=[Optional()])

class Modification(FlaskForm):
    code_faucon = SelectField("falcon_code", choices=[], validators=[Optional()])
    nom_faucon = StringField("nom_faucon", validators=[ ])
    update = TextAreaField("code_faucon", validators=[])    